+++ react/index.tsx
@@ -8,9 +8,8 @@ import React, {
 } from 'react'
 import { injectIntl, InjectedIntlProps } from 'react-intl'
 import marked, { Renderer } from 'marked'
-import { last, test } from 'ramda'
 import escapeHtml from 'escape-html'
-import insane from 'insane'
+import insane from '@vtex/insane'
 import { useCssHandles, CssHandles } from 'vtex.css-handles'
 import { useResponsiveValue } from 'vtex.responsive-values'
 
@@ -124,27 +123,10 @@ const sanitizerConfig = {
     'li',
   ],
   allowedAttributes: {
-    a: ['class', 'href', 'title', 'target'],
-    span: ['class'],
-    p: ['class'],
-    div: ['class'],
-    table: ['class'],
-    thead: ['class'],
-    tbody: ['class'],
-    tr: ['class'],
-    td: ['class'],
-    th: ['class'],
-    img: ['class', 'src', 'title', 'alt'],
+    '*': ['class', 'title'],
+    a: ['href', 'target'],
+    img: ['src', 'alt'],
     iframe: ['frameborder', 'height', 'src', 'width', 'style'],
-    h1: ['class'],
-    h2: ['class'],
-    h3: ['class'],
-    h4: ['class'],
-    h5: ['class'],
-    h6: ['class'],
-    ul: ['class'],
-    ol: ['class'],
-    li: ['class'],
   },
   allowedSchemes: ['http', 'https', 'mailto', 'tel'],
 }
@@ -268,9 +250,7 @@ const RichText: FunctionComponent<Props> = ({
         : href
 
       //clean trailing ? or &
-      const cleanHref = test(/\?|\&/, last(targetRemoved))
-        ? targetRemoved.slice(0, -1)
-        : targetRemoved
+      const cleanHref = targetRemoved.replace(/(\?|&)$/, '')
       const titleAtr = title ? `title="${title}"` : ''
 
       let finalLink = `<a class="${handles.link}" href="${cleanHref}"`
@@ -342,7 +322,7 @@ const RichText: FunctionComponent<Props> = ({
       marked(formatIOMessage({ id: text, intl })),
       sanitizerConfig
     )
-  }, [text, intl, sanitizerConfig, marked, insane, formatIOMessage])
+  }, [text, intl])
 
   return (
     <div