+++ react/typings/insane.d.ts
@@ -7,11 +7,11 @@ declare namespace insane {
   interface IOptions {
     allowedTags?: string[] | boolean
     allowedAttributes?: { [index: string]: string[] } | boolean
-    allowedStyles?:  { [index: string]: { [index: string]: RegExp[] } }
+    allowedStyles?: { [index: string]: { [index: string]: RegExp[] } }
     allowedClasses?: { [index: string]: string[] } | boolean
   }
 }
 
-declare module 'insane' {
+declare module '@vtex/insane' {
   export = insane
 }