+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Replaced `insane` with `@vtex/insane`.
+- Removed `ramda` calls.
 
 ## [0.10.1] - 2020-07-09
 ### Added